backend/requirements/requirements.txt
Django >=4.0.0, <4.1.0
djangorestframework >=3.13.1, <3.14.0
psycopg2-binary>=2.9.0, <2.10.0
drf-spectacular>=0.22.1, <0.23.0
aiohttp >=3.8.1, <3.9.0